basketball-rag/requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
tiktoken>=0.3.0
flask>=2.0.0
python-dotenv>=0.19.0
openpyxl>=3.1.0
argparse>=1.4.0
sqlalchemy>=2.0.0
plotly>=5.13.0  # For visualization if needed